Add files via upload
The program will display
menu contain surgery charges with their prices, pharmacy charges with their prices, and service
charges with their prices based on frequency. For hospital stay, the program will get the number
of days that the patient spent in the hospital from the user and calculate prices based on room
types and food charges. In this program, the user needs to input type of surgery, type of
medication and types of services and then check the days spent in the hospital. The program will
calculate the total of patient’s charges which includes surgery charges, hospital stay, pharmacy
charges and other services charges. The program also will print the total charges when the
patient checks out.